Simulate networks under Exponential Random Graph Models (ERGMs) under local dependence

This function simulates networks under the Exponential Random Graph Model (ERGM) with local dependence with all parameters set according to the estimated model (object). See simulate_bigergm for details of the simulation process

Usage

# S3 method for class 'bigergm'
simulate(
  object,
  nsim = 1,
  seed = NULL,
  ...,
  output = "network",
  control_within = ergm::control.simulate.formula(),
  only_within = FALSE,
  verbose = 0
)

Arguments

object

an object of class bigergm

nsim

number of networks to be randomly drawn from the given distribution on the set of all networks.

seed

seed value (integer) for network simulation.

...

Additional arguments, passed to simulate_formula.

output

Normally character, one of "network" (default), "stats", "edgelist", to determine the output of the function.

control_within

control.simulate.formula object for fine-tuning ERGM simulation of within-block networks.

only_within

If this is TRUE, only within-block networks are simulated.

verbose

If this is TRUE/1, the program will print out additional information about the progress of simulation.

Value

Simulated networks, the output form depends on the parameter output (default is a list of networks).